+++ src/app/[locale]/layout.tsx
@@ -130,7 +130,7 @@ export default async function LocaleLayout({
             </QueryProvider>
             <Script
                 id="fb-pixel"
-                strategy="lazyOnload"
+                strategy="afterInteractive"
                 dangerouslySetInnerHTML={{
                     __html: `
                         !function(f,b,e,v,n,t,s)